refactor(common): split activation precompile module
